New Apps: MotionDescriptors and NetworkProcessing
MotionDescriptors app reinitialized with Trevor Martin's code as of
March 30, 2017. This app is for development of code to extract motion
descriptors following the definitions in Caroline Larboulette and Sylvie
Gibet. 2015. A review of computable expressive descriptors of human
motion. In Proceedings of the 2nd International Workshop on Movement and
Computing (MOCO '15). ACM, New York, NY, USA, 21-28.
DOI=http://dx.doi.org/10.1145/2790994.2790998.

NetworkProcessing app added. This app is for experimentation with
streaming data to the Processing sketches (processing.org).